Fix five rate bugs the audit found
Four are the same class the frame-rate work existed to remove - a whole step
taken once per update rather than once per tick - each in a place the
conversion had not reached:

- The throttle. IncreaseSpeed and DecreaseSpeed still stepped a whole unit
  per update while the roll and pitch beside them on the same class had been
  converted, so holding accelerate reached maximum speed in two-thirds of a
  second at sixty frames a second instead of three.
- Missiles. They are the one thing that steers on every update rather than
  one count in eight, so the acceleration they ask for has to be a rate;
  left as a step a missile reached its top speed in a quarter of the time
  and became far harder to outrun than the game was tuned for. Acceleration
  is a float now, for the same reason velocity was.
- The laser bolt. Drawn for two updates rather than two ticks, which at
  sixty frames a second is a 33ms flicker where it should be about 150ms.
- The docking computer. ApplyAutoDockSpeed nudges the throttle each time it
  runs and it runs every update, so it accelerated and braked four and a
  half times too sharply.

The fifth is from the simulate/compose split rather than the rates. Whether
the screen shows a universe was asked inside the renderer, which after the
split meant asking at compose time - and docking and hyperspace both change
the screen part way through a tick, so the whole view blanked for a frame.
It is asked once now, at the top of the move pass, and remembered: the frame
belongs to the screen the tick started on. The renderer no longer decides
it, which is the right side of the split for it to be on anyway.

All ten baselines are untouched: every one of these is a no-op at the game's
own rate, which is exactly why they survived this long.

+++ src/elite/libs/EliteSharpLib/Ships/IShip.cs
@@ -11,7 +11,7 @@ internal interface IShip : IObject
 {
     public ThreeDModel Model { get; set; }
 
-    public int Acceleration { get; set; }
+    public float Acceleration { get; set; }
 
     public float Bounty { get; set; }
